# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Borsa Batoi ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empresas": {
            "get": {
                "tags": [
                    "empreses"
                ],
                "summary": "Dades de les empreses",
                "description": "Torna les dades de les empreses",
                "operationId": "indexEmpreses",
                "responses": {
                    "200": {
                        "description": "Empreses segons el permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EmpresaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "empreses"
                ],
                "summary": "Guardar empresa",
                "description": "Guardar perfil empresa",
                "operationId": "storeEmpreses",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa correctament donada d'alta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credencials no vàlides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/empresas/{id}": {
            "get": {
                "tags": [
                    "empreses"
                ],
                "summary": "Dades d'una empresa",
                "description": "Torna les dades d'una empresa",
                "operationId": "showEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa si te permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EmpresaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "empreses"
                ],
                "summary": "Modificar empresa",
                "description": "Modificar perfil empresa",
                "operationId": "updateEmpreses",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa correctament donat d'alta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credencials no vàlides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "empreses"
                ],
                "summary": "Esborra Empresa",
                "description": "Torna id de l'empresa esborrada",
                "operationId": "deleteEmpreses",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa esborrasa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlumnoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/menu": {
            "get": {
                "tags": [
                    "menu"
                ],
                "summary": "Menu de l'aplicatiu",
                "description": "Torna les dades dels menu que li correspon a l'usuari",
                "operationId": "indexMenu",
                "responses": {
                    "200": {
                        "description": "Menu de l'aplicatiu segons permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MenuResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "421": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmpresaStoreRequest": {
                "title": "Empresa",
                "description": "Fillable Empresa",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador d'empresa",
                        "type": "integer",
                        "example": "2"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "Nom de l'alumne",
                        "type": "string",
                        "example": "Jose Manuel"
                    },
                    "domicilio": {
                        "title": "domicilio",
                        "description": "Adreça de l'alumne",
                        "type": "string",
                        "example": "C/Cid 14"
                    },
                    "telefono": {
                        "title": "telefono",
                        "description": "Telèfon de l'alumne",
                        "type": "string",
                        "example": "666666666"
                    },
                    "cif": {
                        "title": "cif",
                        "description": "Cif de l'empresa",
                        "type": "string",
                        "example": "A12345678"
                    },
                    "localidad": {
                        "title": "localidad",
                        "description": "Localidad de l'empresa",
                        "type": "string",
                        "example": "Alcoi"
                    },
                    "contacto": {
                        "title": "contacto",
                        "description": "Persona de contacte de l'empresa",
                        "type": "string",
                        "example": "Pepe Botera"
                    },
                    "web": {
                        "title": "web",
                        "description": "Pagina web de l'empresa",
                        "type": "string",
                        "example": "http://www.pepebotera.com"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "Descripció de l'empresa",
                        "type": "string",
                        "example": "Chapuzas a domicilio"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Empresa"
                }
            },
...
            "AlumnoCicloResource": {
                "title": "AlumneCicle",
                "description": "Informació del cicles d'un alumne",
                "properties": {
                    "id_alumno": {
                        "title": "id_alumno",
                        "description": "Identificador d'usuari",
                        "type": "integer",
                        "example": "2"
                    },
                    "id_ciclo": {
                        "title": "id_ciclo",
                        "description": "ID del cicle",
                        "type": "integer",
                        "example": "5"
                    },
                    "any": {
                        "title": "any",
                        "description": "Any acabament del cicle",
                        "type": "string",
                        "example": "2018"
                    },
                    "validado": {
                        "title": "validado",
                        "description": "El cicle ha estat validad per responsable",
                        "type": "boolean",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AlumnoCicloResource"
                }
            },
            "AlumnoResource": {
                "title": "Alumne",
                "description": "Dades d'alumne",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador d'usuari",
                        "type": "integer",
                        "example": "2"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "Nom de l'alumne",
                        "type": "string",
                        "example": "Jose Manuel"
                    },
                    "apellidos": {
                        "title": "apelllidos",
                        "description": "Cognoms de l'alumne",
                        "type": "string",
                        "example": "Miró García"
                    },
                    "domicilio": {
                        "title": "domicilio",
                        "description": "Adreça de l'alumne",
                        "type": "string",
                        "example": "C/Cid 14"
                    },
                    "info": {
                        "title": "info",
                        "description": "Info de l'alumne",
                        "type": "boolean",
                        "example": "1"
                    },
                    "bolsa": {
                        "title": "bolsa",
                        "description": "bolsa",
                        "type": "boolean",
                        "example": "0"
                    },
                    "cv_enlace": {
                        "title": "cv_enlace",
                        "description": "Enllaç al curriculum de l'alumne",
                        "type": "string",
                        "example": "https://www.pepebotera.com"
                    },
                    "telefono": {
                        "title": "telefono",
                        "description": "Telèfon de l'alumne",
                        "type": "string",
                        "example": "666666666"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email de l'alumne",
                        "type": "string",
                        "example": "pepe.botera@gmail.com"
                    },
                    "ciclos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlumnoCicloResource"
                        }
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Hora de creació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Hora de modificació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Alumno"
                }
            },
...
            "EmpresaResource": {
                "title": "Empresa",
                "description": "Dades d'empresa",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador d'empresa",
                        "type": "integer",
                        "example": "6"
                    },
                    "cif": {
                        "title": "cif",
                        "description": "CIF de l'empresa",
                        "type": "string",
                        "example": "B14567890"
                    },
                    "nombre": {
                        "title": "nombre",
                        "description": "Nom de l'empresa",
                        "type": "string",
                        "example": "AITEX"
                    },
                    "domicilio": {
                        "title": "domicilio",
                        "description": "Adreça de l'alumne",
                        "type": "string",
                        "example": "C/Cid 14"
                    },
                    "telefono": {
                        "title": "telefono",
                        "description": "Telèfon de l'empresa",
                        "type": "string",
                        "example": "666666666"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email de l'empresa",
                        "type": "string",
                        "example": "info@aitex.es"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Hora de creació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Hora de modificació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Empresa"
                }
            },
...
            "MenuResource": {
                "title": "Menu",
                "description": "Menu d'usuari",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador de menu",
                        "type": "integer",
                        "example": "6"
                    },
                    "order": {
                        "title": "order",
                        "description": "Ordre del menu",
                        "type": "integer",
                        "example": "4"
                    },
                    "icon": {
                        "title": "icon",
                        "description": "Icona del menu",
                        "type": "string",
                        "example": "account_box"
                    },
                    "text": {
                        "title": "text",
                        "description": "Texte del menu",
                        "type": "string",
                        "example": "Perfil"
                    },
                    "path": {
                        "title": "Path",
                        "description": "Path del menu",
                        "type": "string",
                        "example": "/perfil"
                    },
                    "rol": {
                        "title": "rol",
                        "description": "rol",
                        "type": "integer",
                        "example": "210"
                    },
                    "parent": {
                        "title": "parent",
                        "description": "Pare del menu",
                        "type": "string",
                        "example": "null"
                    },
                    "model": {
                        "title": "model",
                        "description": "???",
                        "type": "boolean",
                        "example": "1"
                    },
                    "active": {
                        "title": "active",
                        "description": "Es Actiu?",
                        "type": "boolean",
                        "example": "1"
                    },
                    "comments": {
                        "type": "string",
                        "example": "null"
                    },
                    "icon_alt": {
                        "title": "icon_at",
                        "description": "Icona alternativa",
                        "type": "string",
                        "example": "null"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Menu"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}